Added yaml data storage functions
The server operator should be able to later change the storage options. Since the amount of data which is saved could get bigger over time, yaml storage allows the data to not being loaded into ram before it actually gets used. While this is useful, it is slow to parse and read data. It is important to cache the loaded data and not check for it every few seconds.