getAccount now calls getAccountAsync in a thread
getAccount is requesting data from an external source, making this in a thread is important to prevent any lag coming from it. There is no not threaded alternative, since there is really no need to make it not threaded. If somebody would use the not threaded function, they would have a bad experience.