The steem account is now loaded async & applied block vector location fix
For some random reason, worlds were not correctly pasted. I really didn't know why this happened but decreasing the vector locations by 0.5 each has helped to fix it. This may come from some random rounding. Sice the first block is always at a 0.5 location, which could make it hard for WorldEdit to decide, where it should be pasted. By changing the values, WorldEdit doesn't has to guess.

This might not be a permanent fix, but it works for now pretty good. If there are any errors, this change is pretty small and can be undone fast.