Added loadSteemWorldData function
The newly added loadSteemWorldData function contains the code which has been in the loadSteemWorld function previousy. Since this data is now also needed in the saving process to check for not changed chunks to prevent saving chunks which haven't changed at all, this is now a separate function.

This removes the need to write it twice, whicch makes this addon more compact.