getSteemContent can now also load older versions of content
In some cases, it is useful to get an older version of content, by adding a blockid, in which the older version is stored, itt will return the older version instead of the new one.

Call it with getSteemContent(Immanuel94,"test",false,200) to get the post with the permlink "test" out of the block number 200. If it doesn't find anything, it will set the metadata result to nothing.

Also added optional async to getSteemContent. If the 3rd parameter is false, the function will run without a thread.