Added getWorldPermission function
The new getWorldPermission function allows to check if the player is allowed to change settings or build in a world. It is pretty simple right now and only allows the direct owner of the world to do anything, but it can be enhanced later.

Using the function instead of directly checking for permissions, improvements can be made easily later. While the permissiontype parameter of the function does not anything yet, it will be used later to give specific permissions later on. Like giving build rights to specific users or interaction permissions.