Added CreateCommentCommand function
This function lets use the /steem command to create comments on Steem. To make sure the player really wants to comment something, the function creates a new random uuid string which has to be entered, thanks to clickable links, this is only a click for players but not possible for maps to trick people into publishing malicious comments.